requirements.txt
colorgram.py==1.2.0
Pillow==9.0.0
